db-mcp-server/requirements.txt
# Database MCP Server Dependencies

# Core MCP framework
fastmcp==2.12.5

# Database drivers
PyMySQL==1.1.2          # MySQL
psycopg2-binary==2.9.9  # PostgreSQL

# Caching
cachetools==6.2.1

# Configuration
python-dotenv==1.1.1

# FastAPI Client Dependencies
fastapi==0.119.0
uvicorn[standard]==0.37.0
httpx==0.28.1
pydantic==2.12.3

# LLM Integration
# nail-client>=1.0.0
google-genai==1.45.0

# SQLAlchemy ORM
sqlalchemy==2.0.44

# Optional: Additional database drivers
# cx_Oracle>=8.3.0      # Oracle
# pymssql>=2.2.0        # SQL Server
